Tests/LibWeb/Text/expected/css/url-style-property-value.txt

url("image.svg")
url(image.svg)
